+++ packages/web-sdk/src/index.ts
@@ -5,7 +5,7 @@ export { defaultEventDomain } from './consts';
 
 export { initializeFaro } from './initialize';
 
-export { getDataFromSafariExtensions, getStackFramesFromError, parseStacktrace, buildStackFrame } from './utils';
+export { getDataFromSafariExtensions, getStackFramesFromError, parseStacktrace, createStacktraceParser, buildStackFrame } from './utils';
 
 export {
   ConsoleInstrumentation,